♻️ refactor: 서비스 초기화 로직을 커스텀 훅으로 분리
main.basic.js의 핵심 서비스 및 DOM 의존 서비스 초기화 로직을
useServiceInitialization.js 커스텀 훅으로 이동하여 구조를 개선함
